CIRC-578: Declared lost item: checkout (failure). Changes regarding comments(2).
